Squashed 'ffmpeg-subtree/' changes from d2b46c1ef7..2d59ca0a66
2d59ca0a66 avcodec: add VMX1 decoder
3b0ed2365c vulkan/h264: set non-existing flag for invalid gaps
1460acc2ac avformat/hlsenc: fix CODECS attribute of H.264
09621fd7d9 avformat/jpegxl_probe: Forward error codes
1ec4553e35 avformat/jpegxl_probe: check length instead of blindly reading
25c937c0e0 avformat/jpegxl_probe: Remove intermediate macro obfuscation around get_bits*()
d78bffbf3d libavutil: Add version bump for new aarch64 cpu flags
18b2ecc247 avcodec/wavarc: add 5elp mode support
024c30aa3b avcodec/exr: remove wrong scaling for 32bit DWA compression
b7e79ba395 avformat/mov: decrease probe score for a random tag
32a749c7a6 avfilter/dnn_backend_openvino: fix log message
6220252f42 configure: add pkg-config check for libopenvino
3a5d95e3fa avfilter/dnn_backend_tf: silence implicit cast warning
b0c0fedcda avfilter/dnn_backend_tf: fix use of uninitialized value
d9f41a343e avfilter/dnn_backend_tf: check TF_OperationOutputType return value
f3495ef4f8 avfilter/dnn_backend_tf: remove unused define
e67af9e7f7 avfilter/vf_dnn_processing: replace magic number by enum value
e7e80e2083 avfilter/vf_dnn_detect: replace magic number by enum value
a887bc4ff1 avfilter/vf_dnn_classify: replace magic number by enum value
016f2f61c3 avfilter/dnn: add log context to ff_get_dnn_module
505c43bb65 avfilter/dnn: refactor ff_get_dnn_module to remove allocation
3f52b7eedc avfilter/dnn: define each backend as a DNNModule
2c4b6976d5 vulkan/hevc: fix 32x32 scaling list indexing
bfd5090628 avfilter/af_rubberband: rewrite EOF handling
1d7a6eaee6 avcodec/proresenc_kostya: set avctx->profile from ctx
eff565dc19 hwcontext_vulkan: tune execution pools
24c4307b80 vulkan_decode: halve execution pool size
5f1be341c2 vulkan: discard dependencies when explicitly waiting for execution
975cd48bb3 vulkan: synchronize access to execution pool fences
697382168d hevcdec: remove redundant bits_used_for_short_term_rps field
d8c7fea8ef lavfi/vf_blend_vulkan: fix leak on error
dae393fe72 libavfilter: add vf_xfade_vulkan
f11515c77f avcodec: add RTV1 decoder
e62824835d lavfi/vf_xfade: rewrite activate inputs handling
0889ebc577 avcodec/noise_bsf: Check for wrapped frames
8d5de4e12f tools/target_dec_fuzzer: Test lowres
402d98c9d4 libavcodec/cuviddec: determine amount of decoded surfaces from within cuvid parser
20584cdd89 avformat/av1: reduce the scope of a variable
a5e45988ac avformat/av1dec: don't return EIO when the input file is truncated
da61ed80fd avformat/av1dec: fix EOF check in Annex-B demuxer
65a1e8ee2c vulkan/hevc: fix scaling lists sizes.
e79686be96 lavc/aarch64: new optimization for 8-bit hevc_qpel_h hevc_qpel_uni_w_hv
15972cce8c lavc/aarch64: new optimization for 8-bit hevc_qpel_uni_w_h
0b7356c1b4 lavc/aarch64: new optimization for 8-bit hevc_pel_uni_w_pixels and qpel_uni_w_v
c76643021e aarch64: Add Windows runtime detection of the dotprod instructions
9b0052200a aarch64: Add Apple runtime detection of dotprod and i8mm using sysctl
493fcde50a aarch64: Add Linux runtime cpu feature detection using HWCAP_CPUID
397cb623c8 aarch64: Add cpu flags for the dotprod and i8mm extensions
fb1b88af77 configure: aarch64: Support assembling the dotprod and i8mm arch extensions
fa11c4c7fa avformat/jpegxl_anim_dec: add animated JPEG XL demuxer
99da411322 avcodec/libjxldec: add animated decode support
141d11cb3c fftools/ffmpeg_mux: make OutputStream.pkt private
730a2221af fftools/ffmpeg_enc: use a private AVPacket instance for encoding
f94957e8f4 fftools/ffmpeg_mux: simplify calling of_output_packet()
96e1325d91 fftools/ffmpeg_mux: use a dedicated packet for BSF output
c803b36b8f fftools/ffmpeg_enc: stop using OutputStream.initialized
568d414074 fftools/ffmpeg: attach filter framerate to frames
9630341073 fftools/ffmpeg: factor out attaching FrameData to a frame
d7781cfb95 fftools/ffmpeg: convert timestamps inside the muxer
cf121592c5 fftools/ffmpeg_mux_init: do not overwrite OutputStream.frame_rate for encoding
5bea8daa78 fftools/ffmpeg_enc: do not guess frame durations from output framerate
2f92650c5d fftools/ffmpeg_enc: inline init_encoder_time_base() into its callers
93e26a4db7 fftools/ffmpeg: handle -enc_time_base -1 during stream creation
58a64e3d54 fftools/ffmpeg_mux_init: only process -enc_time_base if the stream is encoded
2baae6434b fftools/ffmpeg_enc: merge two adjacent video-specific blocks
c8a85d1b2f fftools/ffmpeg_mux_init: do not overwrite OutputStream.frame_rate for streamcopy
ccf219e361 fftools/ffmpeg_enc: avoid breaking exactly integer timestamps in vsync code
c7a05ac117 fftools/ffmpeg_demux: do not set AVCodecContext.framerate
d8c61ba723 fftools/ffmpeg_filter: drop a block disabled since 2012
ad14bdbcfd fftools/ffmpeg_filter: drop a write-only variable
57021a68d9 fftools/ffmpeg_mux: set stream duration after the timebase is certainly known
ba1141d8a9 fftools/ffmpeg_enc: move nb_frames{dup,drop} globals into OutputStream
09af34dc91 tests/fate/ffmpeg: add tests for -max_error_rate
b1c3d81e71 avformat/oggparsetheora: clip duration within 64bit
b168aeb734 avcodec/rka: avoid undefined multiply in  cmode==0
cbe5e480aa avcodec/rka: use 64bit for srate_pad computation
b92af7b64e avcodec/aacenc: add strict bit rate control option
378fb40282 avutil/hwcontext_vulkan: disable multiplane when deriving from cuda
4ef5e7d472 lavc/videotoolboxenc: support additional options
decc776f88 avcodec/videotoolboxenc: add CBP/CHP profile
cab6d7bd71 lavc/videotoolboxenc: use ffmpeg profile constants
2342c05e43 avcodec/magicyuvenc: put some slice work under parallel execution
ba32f28498 doc/examples/transcode: convert timestamps between filtering and encoding
811b0e02be doc/examples/transcode: stop using decoder's AVCodecContext.time_base
828d3b4432 doc/examples/transcode: set packet timebase for decoding
6b8d53f728 avcodec/magicyuvenc: add slice encoding support
4da14c302f avcodec/av1dec: reset the fragment on extradata reading failure
8c6b931f4c avcodec/av1dec: reset the fragment on reading failure
9a9b56c870 avfilter/vf_mix: add fast filtering in case all tmix weights are same
da23151eaf avformat/oggparseflac: use the GetByteContext API

git-subtree-dir: ffmpeg-subtree
git-subtree-split: 2d59ca0a668a0bdf9c83b35fe3ad357f87f8a15d

+++ libavcodec/rka.c
@@ -207,7 +207,7 @@ static int chctx_init(RKAContext *s, ChContext *c,
     c->bprob[0] = s->bprob[0];
     c->bprob[1] = s->bprob[1];
 
-    c->srate_pad = (sample_rate << 13) / 44100 & 0xFFFFFFFCU;
+    c->srate_pad = ((int64_t)sample_rate << 13) / 44100 & 0xFFFFFFFCU;
     c->pos_idx = 1;
 
     for (int i = 0; i < FF_ARRAY_ELEMS(s->bprob[0]); i++)
@@ -732,7 +732,7 @@ static int decode_filter(RKAContext *s, ChContext *ctx, ACoder *ac, int off, uns
                 if (bits == 0) {
                     ctx->buf1[off] = sum + val;
                 } else {
-                    ctx->buf1[off] = (val + (sum >> bits)) * (1 << bits) +
+                    ctx->buf1[off] = (val + (sum >> bits)) * (1U << bits) +
                         (((1U << bits) - 1U) & ctx->buf1[off + -1]);
                 }
                 ctx->buf0[off] = ctx->buf1[off] + ctx->buf0[off + -1];
